+++ packages/common/src/createKmsAccount.ts
@@ -1,8 +1,8 @@
 import { KMSClient } from "@aws-sdk/client-kms";
 import { LocalAccount, hashMessage, hashTypedData, keccak256, serializeTransaction } from "viem";
 import { toAccount } from "viem/accounts";
-import { getEthAddressFromKMS } from "./account/kms";
-import { sign } from "./account/sign";
+import { signWithKms } from "./account/kms/signWithKms";
+import { getAddressFromKMS } from "./account/kms/getAddressFromKms";
 
 export type CreateKmsAccountOptions = {
   keyId: string;
@@ -18,13 +18,13 @@ export async function createKmsAccount({
   keyId,
   client = new KMSClient(),
 }: CreateKmsAccountOptions): Promise<LocalAccount> {
-  const address = await getEthAddressFromKMS({ client, keyId });
+  const address = await getAddressFromKMS({ keyId, client });
 
   const account = toAccount({
     address,
     async signMessage({ message }) {
       const hash = hashMessage(message);
-      const signature = await sign({
+      const signature = await signWithKms({
         client,
         keyId,
         hash,
@@ -36,7 +36,7 @@ export async function createKmsAccount({
     async signTransaction(transaction) {
       const unsignedTx = serializeTransaction(transaction);
       const hash = keccak256(unsignedTx);
-      const signature = await sign({
+      const signature = await signWithKms({
         client,
         keyId,
         hash,
@@ -47,7 +47,7 @@ export async function createKmsAccount({
     },
     async signTypedData(typedData) {
       const hash = hashTypedData(typedData);
-      const signature = await sign({
+      const signature = await signWithKms({
         client,
         keyId,
         hash,
